Turned the mouse position values into doubles.

+++ include/Input.hpp
@@ -10,8 +10,8 @@ namespace wars
 {
     struct MousePos
     {
-        uint32_t x;
-        uint32_t y;
+        double x;
+        double y;
     };
     
     
@@ -29,9 +29,9 @@ namespace wars
         
         bool isMouseButtonDown(int32_t button) { return m_mouseButtons[button]; }
         
-        uint32_t getMouseX() { return m_mousePos.x; }
+        double getMouseX() { return m_mousePos.x; }
         
-        uint32_t getMouseY() { return m_mousePos.y; }
+        double getMouseY() { return m_mousePos.y; }
         
         MousePos getMousePos() { return m_mousePos; }
